add payment, accept, reject and some validation

+++ server/app/Http/Controllers/EventController.php
@@ -2,22 +2,37 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\User;
 use App\Models\Event;
 use App\Models\Narasumber;
-use App\Models\User;
 use Illuminate\Http\Request;
+use App\Models\RegistrasiEvent;
+use Illuminate\Validation\Rule;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Validator;
 
 class EventController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
     public function index()
     {
         $events = Event::with([
-            'narasumber' => function ($query) {
+            'nid1' => function ($query) {
+                $query->select('id', 'name', 'keterangan', 'image');
+            },
+            'nid2' => function ($query) {
+                $query->select('id', 'name', 'keterangan', 'image');
+            },
+            'nid3' => function ($query) {
+                $query->select('id', 'name', 'keterangan', 'image');
+            },
+            'nid4' => function ($query) {
+                $query->select('id', 'name', 'keterangan', 'image');
+            },
+            'nid5' => function ($query) {
+                $query->select('id', 'name', 'keterangan', 'image');
+            },
+            'nid6' => function ($query) {
                 $query->select('id', 'name', 'keterangan', 'image');
             },
             'user' => function ($query) {
@@ -25,22 +40,16 @@ class EventController extends Controller
             }
         ])->get();
 
+        $eventsData = $events->map(function ($event) {
+            $event['price'] = formatPrice($event['price']);
+            return $event;
+        });
+
         return response()->json([
-            "data" => $events,
+            "data" => $eventsData,
         ], 200);
     }
 
-    /**
-     * Show the form for creating a new resource.
-     */
-    public function create()
-    {
-        //
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     */
     public function store(Request $request)
     {
         $validatedData = Validator::make($request->all(), [
@@ -52,7 +61,7 @@ class EventController extends Controller
             "status" => ['required'],
             "event_date" => ['required'],
             "lokasi" => ['required'],
-            "narasumber_id" => ['required'],
+            "nid_1" => ['required'],
         ]);
 
         if ($validatedData->fails()) {
@@ -63,21 +72,27 @@ class EventController extends Controller
         }
 
         $slug = strtolower(slugify($request->input('title')));
+        $image = base64_encode($request->image);
         $user = Auth::user();
 
         $event = Event::create([
             'slug' => $slug,
             'title' => $request->input('title'),
             'keterangan' => $request->keterangan,
-            'image' => $request->image,
+            'image' => $image,
             'price' => $request->price,
             'quota' => $request->quota,
-            'status' => $request->status,
+            'status' => 'upcoming',
             'publish_date' => date('Y-m-d'),
             'event_date' => $request->event_date,
             'published' => 0,
             'lokasi' => $request->lokasi,
-            'narasumber_id' => $request->narasumber_id,
+            'nid_1' => $request->nid_1,
+            'nid_2' => $request->nid_2,
+            'nid_3' => $request->nid_3,
+            'nid_4' => $request->nid_4,
+            'nid_5' => $request->nid_5,
+            'nid_6' => $request->nid_6,
             'user_id' => $user->id,
         ]);
 
@@ -87,9 +102,6 @@ class EventController extends Controller
         ], 201);
     }
 
-    /**
-     * Display the specified resource.
-     */
     public function show(Request $request, Event $event)
     {
         $event = Event::find($request->id);
@@ -100,17 +112,72 @@ class EventController extends Controller
             ], 404);
         }
 
-        $narasumber = Narasumber::find($event->narasumber_id);
+        $eventRegistrationCount = RegistrasiEvent::where('event_id', $event->id)->count();
+        $quotaNow = $event->quota - $eventRegistrationCount;
+
+        $nid_1 = Narasumber::find($event->nid_1);
+        $nid_2 = Narasumber::find($event->nid_2);
+        $nid_3 = Narasumber::find($event->nid_3);
+        $nid_4 = Narasumber::find($event->nid_4);
+        $nid_5 = Narasumber::find($event->nid_5);
+        $nid_6 = Narasumber::find($event->nid_6);
         $user = User::find($event->user_id);
 
         return response()->json([
-            "data" => $event,
-            "narasumber" => [
-                'id' => $narasumber->id,
-                'name' => $narasumber->name,
-                'keterangan' => $narasumber->keterangan,
-                'image' => $narasumber->image,
+            "data" => [
+                'id' => $event->id,
+                'slug' => $event->slug,
+                'title' => $event->title,
+                'keterangan' => $event->keterangan,
+                'image' => $event->image,
+                'price' => formatPrice($event->price),
+                'quota' => $event->quota,
+                'quota_now' => $quotaNow,
+                'status' => $event->status,
+                'user_id' => $event->user_id,
+                'publish_date' => $event->publish_date,
+                'event_date' => $event->event_date,
+                'published' => $event->published,
+                'lokasi' => $event->lokasi,
+                'created_at' => $event->created_at,
+                'updated_at' => $event->updated_at
             ],
+            "nid_1" => [
+                'id' => $nid_1->id,
+                'name' => $nid_1->name,
+                'keterangan' => $nid_1->keterangan,
+                'image' => $nid_1->image,
+            ],
+            "nid_2" => $nid_2 ? [
+                'id' => $nid_2->id,
+                'name' => $nid_2->name,
+                'keterangan' => $nid_2->keterangan,
+                'image' => $nid_2->image,
+            ] : null,
+            "nid_3" => $nid_3 ? [
+                'id' => $nid_3->id,
+                'name' => $nid_3->name,
+                'keterangan' => $nid_3->keterangan,
+                'image' => $nid_3->image,
+            ] : null,
+            "nid_4" => $nid_4 ? [
+                'id' => $nid_4->id,
+                'name' => $nid_4->name,
+                'keterangan' => $nid_4->keterangan,
+                'image' => $nid_4->image,
+            ] : null,
+            "nid_5" => $nid_5 ? [
+                'id' => $nid_5->id,
+                'name' => $nid_5->name,
+                'keterangan' => $nid_5->keterangan,
+                'image' => $nid_5->image,
+            ] : null,
+            "nid_6" => $nid_6 ? [
+                'id' => $nid_6->id,
+                'name' => $nid_6->name,
+                'keterangan' => $nid_6->keterangan,
+                'image' => $nid_6->image,
+            ] : null,
             "users" => [
                 'id' => $user->id,
                 'name' => $user->name
@@ -118,17 +185,6 @@ class EventController extends Controller
         ], 200);
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit(Event $event)
-    {
-        //
-    }
-
-    /**
-     * Update the specified resource in storage.
-     */
     public function publish(Request $request, Event $event)
     {
         $event = Event::find($request->id);
@@ -146,7 +202,7 @@ class EventController extends Controller
         $event = Event::find($request->id);
 
         return response()->json([
-            "message" => "Event updated successfully",
+            "message" => "Event published successfully",
             "data" => [
                 'id' => $event->id,
                 'published' => $event->published,
@@ -156,13 +212,17 @@ class EventController extends Controller
         ], 200);
     }
 
-    /**
-     * Update the specified resource in storage.
-     */
     public function update(Request $request, Event $event)
     {
+        $event = Event::find($request->id);
+        if (!$event) {
+            return response()->json([
+                "error" => "Event not found"
+            ], 404);
+        }
+
         $validatedData = Validator::make($request->all(), [
-            "title" => ['required', 'unique:events,title'],
+            "title" => ['required', Rule::unique('events')->ignore($event->id)],
             "keterangan" => ['required'],
             "image" => ['required'],
             "price" => ['required'],
@@ -170,15 +230,8 @@ class EventController extends Controller
             "status" => ['required'],
             "event_date" => ['required'],
             "lokasi" => ['required'],
-            "narasumber_id" => ['required'],
+            "nid_1" => ['required'],
         ]);
-
-        $event = Event::find($request->id);
-        if (!$event) {
-            return response()->json([
-                "error" => "Event not found"
-            ], 404);
-        }
 
         if ($validatedData->fails()) {
             return response()->json([
@@ -188,21 +241,25 @@ class EventController extends Controller
         }
 
         $slug = strtolower(slugify($request->input('title')));
+        $image = base64_encode($request->image);
         $user = Auth::user();
 
         Event::where('id', $request->id)->update([
             'slug' => $slug,
             'title' => $request->input('title'),
             'keterangan' => $request->keterangan,
-            'image' => $request->image,
+            'image' => $image,
             'price' => $request->price,
             'quota' => $request->quota,
             'status' => $request->status,
-            'publish_date' => date('Y-m-d'),
             'event_date' => $request->event_date,
-            'published' => 0,
             'lokasi' => $request->lokasi,
-            'narasumber_id' => $request->narasumber_id,
+            'nid_1' => $request->nid_1,
+            'nid_2' => $request->nid_2,
+            'nid_3' => $request->nid_3,
+            'nid_4' => $request->nid_4,
+            'nid_5' => $request->nid_5,
+            'nid_6' => $request->nid_6,
             'user_id' => $user->id,
         ]);
 
@@ -214,22 +271,70 @@ class EventController extends Controller
         ], 200);
     }
 
-    /**
-     * Remove the specified resource from storage.
-     */
+    public function tersediaEvent(string $id)
+    {
+        $eventExist = Event::find($id);
+
+        if (!$eventExist) {
+            return response()->json([
+                "error" => "Event not found"
+            ], 422);
+        }
+
+        Event::where('id', $id)->update([
+            'status' => 'available'
+        ]);
+
+        return response()->json([
+            "message" => "Event available"
+        ], 200);
+    }
+
+    public function selesaiEvent(string $id)
+    {
+        $eventExist = Event::find($id);
+
+        if (!$eventExist) {
+            return response()->json([
+                "error" => "Event not found"
+            ], 422);
+        }
+
+        $registrasiTanpaPembayaran = RegistrasiEvent::where('event_id', $id)->whereDoesntHave('pembayaran')->get();
+        foreach ($registrasiTanpaPembayaran as $registrasi) {
+            $registrasi->delete();
+        }
+
+        Event::where('id', $id)->update([
+            'status' => 'unavailable'
+        ]);
+
+        return response()->json([
+            "message" => "Event selesai"
+        ], 200);
+    }
+
     public function destroy(Request $request, Event $event)
     {
         $event = Event::find($request->id);
+
         if (!$event) {
             return response()->json([
                 "error" => "Event not found"
             ], 404);
         }
 
-        Event::where('id', $request->id)->delete();
+        $registrasiEventExists = RegistrasiEvent::where('event_id', $event->id)->first();
 
-        return response()->json([
-            "data" => "ok"
-        ], 200);
+        if ($registrasiEventExists) {
+            return response()->json([
+                "error" => "Tidak dapat menghapus event karena masih ada registrasi event terkait"
+            ], 422);
+        } else {
+            Event::where('id', $request->id)->delete();
+            return response()->json([
+                "message" => "Event berhasil dihapus"
+            ], 200);
+        }
     }
 }
